factor out store/shuffleable
